% 
\docType{data}
\name{AlgaeTV}
\alias{AlgaeTV}
\title{Data theoretical values for Ecological Evaluation Index}
\format{Use data
	\describe{
		\item{Taxa}{species or genus}
		\item{Group}{group ESG}
		\item{A1...D3}{replicas}
	}
}
\source{
	http://eei.gr/index.html
}
\usage{
AlgaeTV
}
\description{
	Marine benthic macrophytes 
}
\keyword{datasets}
