% 
\name{eei}
\alias{eei}
\title{Ecological Evaluation Index}
\usage{
eei(x, y, type = c("esi", "hyp", "eeic"), k = c(0.503, 0.954, -0.204, -0.998, 0.355, -0.109))
}
\arguments{
	\item{x}{ESGI value }
	\item{y}{ESGII value}
	\item{type}{type of indice}
	\item{k}{vector with a,b,d,c,e and f coefficients}
}
\description{
	A biotic index implementation.	
}
