% 
\name{eei_hyp}
\alias{eei_hyp}
\title{EEI Hiperbolic Function}
\usage{
eei_hyp(x, y, k = c(a = 0.503, b = 0.954, c = -0.204, d = -0.998, e = 0.355, f = -0.109))
}
\arguments{
	\item{x}{ESGI value }
	\item{y}{ESGII value}
	\item{k}{vector with a,b,d,c,e and f coefficients of the hyperbola}
}
\description{
	A biotic index implementation.	
}
