% 
\name{eeic}
\alias{eeic}
\title{Ecological Evaluation Index}
\usage{
eeic(x, y, k = c(a = 0.4680, b = 1.2088, c = -0.3583, d = -1.1289, e = 0.5129, f = -0.1869))
}
\arguments{
	\item{x}{ESGI value }
	\item{y}{ESGII value}
	\item{k}{vector with a,b,d,c,e and f coefficients of the hyperbola}
}
\description{
	A biotic index implementation.	
}
