%
\docType{data}
\name{EEImelt}
\alias{EEImelt}
\title{Data for Ecological Evaluation Index}
\format{use data
	\describe{
		\item{specie}{species}
		\item{Group}{group}
		\item{Site}{Local}
		\item{Replica}{Local}
		\item{Coverage}{coveage percentual}
	}
}
\source{
	http://eei.gr/index.html
}
\usage{
EEI
}
\description{
	Marine benthic macrophytes 
}
\keyword{datasets}
