% 
\name{eqr}
\alias{eqr}
\title{Ecological Quality Ratios}
\usage{
eqr(x, rc = 10)
}
\arguments{
	\item{x}{EEIc values}
	\item{rc}{constant}
}
\description{
	To ensure comparability, the EEI-c values were transformed into Ecological Quality Ratios (EQR, i.e. the ratio between the value of the observed biological parameter for a given surface water body and its expected value under the reference conditions)
}
