% 
\name{gsn}
\alias{gsn}
\title{Generate Sequence Number}
\usage{
gsn(rel = NULL, num = NULL, status = NULL, from = 0, to = 100, by = .1)
}
\arguments{
	\item{rel}{data.frame with logic relaction}
	\item{num}{status class numeric value}
	\item{status}{status class catgorical}
	\item{from}{initial value}
	\item{to}{final value}
	\item{by}{interval}
}
\description{
	To ensure comparability, the EEI-c values were transformed into Ecological Quality Ratios (EQR, i.e. the ratio between the value of the observed biological parameter for a given surface water body and its expected value under the reference conditions)
}
