% 
\name{quality_map}
\alias{quality_map}
\title{Ecological Evaluation Index}
\usage{
quality_map(dta, lng = "lng", lat = "lat", esc = "ESC", idl = "")
}
\arguments{
	\item{dta}{data}
	\item{lng}{longitude}
	\item{lat}{latitude}
	\item{esc}{name esc}
	\item{idl}{group}
}
\description{
   Map quality	
}
