% 
\name{desg}
\alias{desg}
\title{Ecological Status Group}
\usage{
desg(dta, coverage, group, place)
}
\arguments{
	\item{dta}{data.table}
	\item{coverage}{name coverage}
	\item{group}{name group ESG}
	\item{place}{local}
}
\description{
The late-successional group with low growth rates and long life histories (Ecological Status Group I, mostly K-selection) and the opportunistic group with high growth rates and short life histories (ESG II, mostly r-selection).
}
