% 
\name{desg}
\alias{desg}
\title{Ecological Status Group}
\usage{
desg(dt, cove, group, ia = NULL, ib = NULL, ic = NULL, iia = NULL, iib = NULL, iic = NULL, site = NULL, replica = NULL)
}
\arguments{
	\item{dt}{data.table}
	\item{cove}{name coverage}
	\item{group}{name group}
	\item{ia}{group IA}
	\item{ib}{group IB}
	\item{ic}{group IC}
	\item{iia}{group IIA}
	\item{iib}{group IIB}
	\item{iic}{group IIC}
	\item{site}{local}
	\item{replica}{replica}
}
\description{
The late-successional group with low growth rates and long life histories (Ecological Status Group I, mostly K-selection) and the opportunistic group with high growth rates and short life histories (ESG II, mostly r-selection).
}
