% 
\docType{data}
\name{EEI}
\alias{EEI}
\title{Data for Ecological Evaluation Index}
\format{Use data
	\describe{
		\item{specie}{species}
		\item{Group}{group}
		\item{A1...D3}{replicas}
	}
}
\source{
	http://eei.gr/index.html
}
\usage{
EEI
}
\description{
	Marine benthic macrophytes 
}
\keyword{datasets}
