% 
\name{eeic}
\alias{eeic}
\title{Ecological Evaluation Index}
\usage{
eeic(x, y, k = c(a = 0.50339, b = 0.95404, c = -0.20394, d = -0.99804, e = 0.35476, f = -0.10909))
}
\arguments{
	\item{x}{ESGI value }
	\item{y}{ESGII value}
	\item{k}{vector with a,b,d,c,e and f coefficients of the hyperbola}
}
\description{
	A biotic index implementation.	
}
