% 
\name{eeic}
\alias{eeic}
\title{Ecological Evaluation Index}
\usage{
eeic(x, y, a = 0.468, b = 1.2088, c = -0.3583, d = -1.1289, e = 0.5129, f = -0.1869)
}
\arguments{
	\item{x}{ESGI value }
	\item{y}{ESGII value}
	\item{a}{coefficients of the hyperbola}
	\item{b}{coefficients of the hyperbola}
	\item{c}{coefficients of the hyperbola}
	\item{d}{coefficients of the hyperbola}
	\item{e}{coefficients of the hyperbola}
	\item{f}{coefficients of the hyperbola}
}
\description{
	A biotic index for the implementation of WFD in rocky coastal and sedimentary.	
}
