% 
\name{esg}
\alias{esg}
\title{Ecological Status Group}
\usage{
esg(a = 0, b = 0, c = 0, k1 = 1, k2 = 0.8, k3 = 0.6, w1 = 0.8, w2 = 1, w3 = 1,
	type = c("1", "2"))
}
\arguments{
	\item{a}{coefficients of the ESG}
	\item{b}{coefficients of the ESG}
	\item{c}{coefficients of the ESG}
	\item{k1}{constant of the ESG}
	\item{k2}{constant of the ESG}
	\item{k3}{constant of the ESG}
	\item{w1}{constant of the ESG}
	\item{w2}{constant of the ESG}
	\item{w3}{constant of the ESG}
	\item{type}{predefined categories of 1-ESG I and 2-ESG II}
}
\description{
the late-successional group with low growth rates and long life histories (Ecological Status Group I, mostly K-selection) and the opportunistic group with high growth rates and short life histories (ESG II, mostly r-selection).
}
