% 
\name{group_esg}
\alias{group_esg}
\title{Ecological Status Group}
\usage{
group_esg(file_dta, rows, sheet, class_esg = NULL, xlsx = FALSE, open.xlsx = FALSE, out.dir = NULL)
}
\arguments{
	\item{file_dta}{file xlsx}
	\item{rows}{line rows}
	\item{sheet}{sheet}
	\item{class_esg}{group esg}
	\item{xlsx}{out file xlsx}
	\item{open.xlsx}{open file xlsx}
	\item{out.dir}{directory}
}
\description{
The late-successional group with low growth rates and long life histories (Ecological Status Group I, mostly K-selection) and the opportunistic group with high growth rates and short life histories (ESG II, mostly r-selection).
}
