% 
\name{gsn}
\alias{gsn}
\title{Generate Sequence Number}
\usage{
gsn(rv = TRUE, from = 0, to = 100, by = .5)
}
\arguments{
	\item{rv}{data.table or logic use default}
	\item{from}{initial value}
	\item{to}{final value}
	\item{by}{interval}
}
\description{
	To ensure comparability, the EEI-c values were transformed into Ecological Quality Ratios (EQR, i.e. the ratio between the value of the observed biological parameter for a given surface water body and its expected value under the reference conditions)
}
