% 
\name{index}
\alias{index}
\title{Index EEIc - EQR - ESC}
\usage{
index(dta, coverage, group, by, mean_by = NULL, k = c(a = 0.503, b = 0.954, c = -0.204, d = -0.998, e = 0.355, f = -0.109))
}
\description{
calculates the EEIc, EQR and ESC indices.
}
\arguments{
	\item{dta}{data.table}
	\item{coverage}{name coverage}
	\item{group}{name group ESG}
	\item{by}{vector}
	\item{mean_by}{mean by}
	\item{k}{constant for EEI}
}

